+++ src/modules/inc/gps_module.h
@@ -1,0 +1,4 @@
+#ifndef CUBESAT_CORE_MODULES_GPS_MODULE_H
+#define CUBESAT_CORE_MODULES_GPS_MODULE_H
+
+#endif
