+++ src/modules/inc/mission_module.h
@@ -3,7 +3,7 @@
 
 // Project headers
 #include "client_api.h"
-#include "subsystem_module.h"
+#include "subsysmod.h"
 
 // Standard C libraries
 #include <stdio.h>
